--- a/src/components.jl
+++ b/src/components.jl
@@ -529,7 +529,7 @@ function series_annotations_shapes!(series::Series, scaletype::Symbol = :pixels)
     # end
 
     # @show msw msh
-    if anns != nothing && anns.baseshape != nothing
+    if anns !== nothing && anns.baseshape !== nothing
         # we use baseshape to overwrite the markershape attribute
         # with a list of custom shapes for each
         msw,msh = anns.scalefactor