Should decrease compile times