@ChrisRackauckas 2019-08-17
From

```
julia> @time using Plots
  5.215902 seconds (10.59 M allocations: 580.846 MiB, 5.43% gc time)

julia> @time plot(rand(4,4))
 16.619736 seconds (49.28 M allocations: 2.408 GiB, 8.68% gc time)
```

to

```
  julia> @time using Plots
    4.865916 seconds (10.48 M allocations: 575.247 MiB, 4.54% gc time)
  
  julia> @time plot(rand(4,4))
   15.923668 seconds (49.26 M allocations: 2.407 GiB, 8.65% gc time)
```