@yha 2019-08-18
I don't see performance gains here (Julia 1.1.1, Windows 10).
Before:
```
julia> @time using Plots
  3.741940 seconds (7.43 M allocations: 434.046 MiB, 5.01% gc time)

julia> @time plot(rand(4,4))
 13.814570 seconds (49.98 M allocations: 2.436 GiB, 9.78% gc time)
```
After:
```
julia> @time using Plots
  3.788572 seconds (7.44 M allocations: 434.568 MiB, 4.92% gc time)

julia> @time plot(rand(4,4))
 13.691652 seconds (50.00 M allocations: 2.437 GiB, 9.97% gc time)
```
But it seems harmless.