@ChrisRackauckas 2019-08-18
Yeah, it's at least harmless and the right direction. There must be other stuff going on, but since we've seen a lot of libraries get a speedup I thought we might as well do it here as well.